Move navigation into UI folder. Decouple it from Redux.
